refactor: extract theme logic to hook
