+++ src/components/ThemeToggle/index.js
@@ -1,37 +1,11 @@
-import React, { useState, useEffect } from 'react';
+import React from 'react';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import { faSun, faMoon } from '@fortawesome/free-solid-svg-icons';
+import useTheme from '../../hooks/useTheme';
 import './index.scss';
 
 const ThemeToggle = ({ className = '', fixed = true }) => {
-    const [isDark, setIsDark] = useState(false);
-
-    useEffect(() => {
-        // Check for saved theme preference or default to light
-        const savedTheme = localStorage.getItem('theme');
-        const mql = window.matchMedia
-            ? window.matchMedia('(prefers-color-scheme: dark)')
-            : null;
-        const prefersDark = mql ? mql.matches : false;
-
-        if (savedTheme === 'dark' || (!savedTheme && prefersDark)) {
-            setIsDark(true);
-            document.documentElement.setAttribute('data-theme', 'dark');
-        }
-    }, []);
-
-    const toggleTheme = () => {
-        const newTheme = !isDark;
-        setIsDark(newTheme);
-
-        if (newTheme) {
-            document.documentElement.setAttribute('data-theme', 'dark');
-            localStorage.setItem('theme', 'dark');
-        } else {
-            document.documentElement.removeAttribute('data-theme');
-            localStorage.setItem('theme', 'light');
-        }
-    };
+    const { isDark, toggleTheme } = useTheme();
 
     const classes = `theme-toggle btn btn--ghost btn--icon ${
         fixed ? 'theme-toggle--fixed' : ''
